// src/Types/Component.ts
export type ColorVariant = 'primary' | 'red' | 'neutral' | 'yellow'

export type ColorWeight = 100 | 200 | 300 | 400 | 500 | 600

export type TypographySize =
  | '2xs'
  | 'xs'
  | 'sm'
  | 'md'
  | 'lg'
  | 'xl'
  | '2xl'
  | '3xl'
  | '4xl'
  | '5xl'
  | '6xl'

export type TypographyType = 'text'

export type TypographyWeight = 'regular' | 'medium' | 'bold'

export type ValueLabelOption = { value: string | number; label: string }

export type ValueLabelOptions = Array<ValueLabelOption>